Remove new deferGraphConstraint option to work around bug.
UpdateVisitSummaryTask can take inputs that include the {tract}
dimension, which can impose an undesirable spatial constraint on its
{visit, detector} inputs when only one tract (or a few disjoint tracts)
are being processed.  Using deferGraphConstraint here sets the default
behavior of the pipetask --dataset-query-constraint option such that
users don't have to use it to get the right behavior for this task (and
other tasks being run with it).

I'm also using deferGraphConstraint on many other connections here as
an optimization, which is something we could start doing in more tasks,
whenever we can be reasonably confident that some other dataset's
existence (here visitSummary) will provide an equivalent constraint.