Add refcat loading for bright star marks.
This has not been tested.  It will definitely need config overrides
to "config.connections.ref_cat" to work:

- "cal_ref_cat" for DC2 data
- "ps1_pv3_3pi_20170110" or "gaia_dr3_20230707" for real data

Right now the reference catalog is being passed to run() as an
lsst.afw.table.SimpleCatalog, which is probably not what anyone wants,
but it's what we start off with.  Calling .asAstropy() on it and
passing that to run() would be one alternative, and calling
.extract(<columns>) to get a dict of numpy array columns would be
another.