Clear single-frame streak mask if it exists already
Before detecting streaks on the difference warps, clear any existing
mask using the same name. Any existing mask has been PSF-matched,
is not a replacement for detecting streaks on the differences, and
should not be used in the artifact rejection loop.

To clip pixels based on the single-frame streak masks, add that mask
name to the badMaskPlanes config parameter.

Streaks detected on the warp differences are expected to be better
in that they are straight, cross multiple detectors, and are not competing
against the background of static sources