Update coadd units based on config
For the short period before retiring the functionality of
`do_scale_zero_point`, we support making cell-coadds with the new
version of the code from both the old and the new warps. In this
interim, we use this boolean config field to indicate whether we are
running this with old warp (True) or not (False).